+++ lua/mini/files.lua
@@ -20,6 +20,7 @@
 ---     - Filter/prefix/sort of file system entries.
 ---     - Mappings used for common explorer actions.
 ---     - UI options: whether to show preview of file/directory under cursor, etc.
+---     - Bookmarks for quicker navigation.
 ---
 --- What it doesn't do:
 --- - Try to be replacement of system file explorer. It is mostly designed to
@@ -157,6 +158,10 @@
 ---  |-------------|------|------------------------------------------------|
 ---  | Go out plus |  H   | Focus on parent directory plus extra action    |
 ---  |-------------|------|------------------------------------------------|
+---  | Go to mark  |  '   | Jump to bookmark (waits for single key id)     |
+---  |-------------|------|------------------------------------------------|
+---  | Set mark    |  m   | Set bookmark (waits for single key id)         |
+---  |-------------|------|------------------------------------------------|
 ---  | Reset       | <BS> | Reset current explorer                         |
 ---  |-------------|------|------------------------------------------------|
 ---  | Reveal cwd  |  @   | Reset current current working directory        |
@@ -181,6 +186,10 @@
 ---
 --- - "Go out plus" is regular "Go out" but trims right part of branch.
 ---
+--- - "Set mark" and "Go to mark" both wait for user to press a single character
+---   of a bookmark id. Example: `ma` sets directory path of focused window as
+---   bookmark "a"; `'a` jumps (sets as whole branch) to bookmark "a".
+---
 --- - "Reset" focuses only on "anchor" directory (the one used to open current
 ---   explorer) and resets all stored directory cursor positions.
 ---
@@ -188,7 +197,7 @@
 ---   If it is not an ancestor of the current branch, nothing is done.
 ---
 --- - "Show help" results into new window with helpful information about current
----   explorer. Press `q` to close it.
+---   explorer (like buffer mappings and bookmarks). Press `q` to close it.
 ---
 --- - "Synchronize" parses user edits in directory buffers, applies them (after
 ---   confirmation), and updates all directory buffers with the most relevant
@@ -504,6 +513,22 @@
 ---     end,
 ---   })
 --- <
+--- # Set custom bookmarks ~
+---
+--- Use |MiniFiles.set_bookmark()| inside `MiniFilesExplorerOpen` event: >lua
+---
+---   local set_mark = function(id, path, desc)
+---     MiniFiles.set_bookmark(id, path, { desc = desc })
+---   end
+---   vim.api.nvim_create_autocmd('User', {
+---     pattern = 'MiniFilesExplorerOpen',
+---     callback = function()
+---       set_mark('c', vim.fn.stdpath('config'), 'Config') -- path
+---       set_mark('w', vim.fn.getcwd, 'Working directory') -- callable
+---       set_mark('~', '~', 'Home directory')
+---     end,
+---   })
+--- <
 ---@tag MiniFiles-examples
 
 ---@diagnostic disable:luadoc-miss-type-name
@@ -643,6 +668,8 @@ MiniFiles.config = {
     go_in_plus  = 'L',
     go_out      = 'h',
     go_out_plus = 'H',
+    mark_goto   = "'",
+    mark_set    = 'm',
     reset       = '<BS>',
     reveal_cwd  = '@',
     show_help   = 'g?',
@@ -965,7 +992,7 @@ MiniFiles.show_help = function()
   local buf_id = vim.api.nvim_get_current_buf()
   if not H.is_opened_buffer(buf_id) then return end
 
-  H.explorer_show_help(buf_id, vim.api.nvim_get_current_win())
+  H.explorer_show_help(explorer, buf_id, vim.api.nvim_get_current_win())
 end
 
 --- Get file system entry data
@@ -992,6 +1019,8 @@ end
 ---@return table|nil Table with explorer state data or `nil` if no active explorer.
 ---   State data is a table with the following fields:
 ---   - <anchor> `(string)` - anchor directory path (see |MiniFiles.open()|).
+---   - <bookmarks> `(table)` - map from bookmark id (single character) to its data:
+---     table with <path> and <desc> fields (see |MiniFiles.set_bookmark()|).
 ---   - <branch> `(table)` - array of nested paths for currently opened branch.
 ---   - <depth_focus> `(number)` - an index in <branch> for currently focused path.
 ---   - <target_window> `(number)` - identifier of target window.
@@ -999,8 +1028,9 @@ end
 ---     Each element is a table with <win_id> (window identifier) and <path> (path
 ---     shown in the window) fields.
 ---
----@seealso - |MiniFiles.set_target_window()|
+---@seealso - |MiniFiles.set_bookmark()|
 --- - |MiniFiles.set_branch()|
+--- - |MiniFiles.set_target_window()|
 MiniFiles.get_explorer_state = function()
   local explorer = H.explorer_get()
   if explorer == nil then return end
@@ -1015,6 +1045,7 @@ MiniFiles.get_explorer_state = function()
 
   return {
     anchor = explorer.anchor,
+    bookmarks = vim.deepcopy(explorer.bookmarks),
     branch = vim.deepcopy(explorer.branch),
     depth_focus = explorer.depth_focus,
     target_window = explorer.target_window,
@@ -1088,6 +1119,27 @@ MiniFiles.set_branch = function(branch, opts)
   -- way is not really feasible, as it requires knowing cursor at deepest path,
   -- which might not yet be set before first refresh.
   H.explorer_refresh(explorer)
+end
+
+--- Set bookmark
+---
+---@param id string Single character bookmark id.
+---@param path string|function Path of a present on disk directory to set as
+---   a bookmark's path. If callable, should return such path.
+---@param opts table|nil Options. Possible fields:
+---   - <desc> `(string)` - bookmark description (used in help window).
+MiniFiles.set_bookmark = function(id, path, opts)
+  local explorer = H.explorer_get()
+  if explorer == nil then return end
+
+  if not (type(id) == 'string' and id:len() == 1) then H.error('Bookmark id should be single character') end
+  local is_valid_path = vim.is_callable(path)
+    or (type(path) == 'string' and H.fs_get_type(vim.fn.expand(path)) == 'directory')
+  if not is_valid_path then H.error('Bookmark path should be a valid path to directory or a callable.') end
+  opts = opts or {}
+  if not (opts.desc == nil or type(opts.desc) == 'string') then H.error('Bookmark description should be string') end
+
+  explorer.bookmarks[id] = { path = path, desc = opts.desc }
 end
 
 --- Get latest used anchor path
@@ -1231,6 +1283,8 @@ H.setup_config = function(config)
     ['mappings.go_in_plus'] = { config.mappings.go_in_plus, 'string' },
     ['mappings.go_out'] = { config.mappings.go_out, 'string' },
     ['mappings.go_out_plus'] = { config.mappings.go_out_plus, 'string' },
+    ['mappings.mark_goto'] = { config.mappings.mark_goto, 'string' },
+    ['mappings.mark_set'] = { config.mappings.mark_set, 'string' },
     ['mappings.reset'] = { config.mappings.reset, 'string' },
     ['mappings.reveal_cwd'] = { config.mappings.reveal_cwd, 'string' },
     ['mappings.show_help'] = { config.mappings.show_help, 'string' },
@@ -1328,6 +1382,8 @@ end
 -- Explorers ------------------------------------------------------------------
 ---@class Explorer
 ---
+---@field bookmarks table Map from single characters to bookmark data: table
+---   with <path> and <desc> fields.
 ---@field branch table Array of absolute directory paths from parent to child.
 ---   Its ids are called depth.
 ---@field depth_focus number Depth to focus.
@@ -1353,6 +1409,7 @@ H.explorer_new = function(path)
     windows = {},
     anchor = path,
     target_window = vim.api.nvim_get_current_win(),
+    bookmarks = {},
     opts = {},
   }
 end
@@ -1809,7 +1866,7 @@ H.explorer_trim_branch_left = function(explorer)
   return explorer
 end
 
-H.explorer_show_help = function(explorer_buf_id, explorer_win_id)
+H.explorer_show_help = function(explorer, explorer_buf_id, explorer_win_id)
   -- Compute lines
   local buf_mappings = vim.api.nvim_buf_get_keymap(explorer_buf_id, 'n')
   local map_data, desc_width = {}, 0
@@ -1829,6 +1886,20 @@ H.explorer_show_help = function(explorer_buf_id, explorer_win_id)
     table.insert(lines, string.format(map_format, desc, map_data[desc]))
   end
   table.insert(lines, '')
+
+  local bookmark_ids = vim.tbl_keys(explorer.bookmarks)
+  if #bookmark_ids > 0 then
+    table.insert(lines, 'Bookmarks:')
+    table.insert(lines, '')
+    table.sort(bookmark_ids)
+    for _, id in ipairs(bookmark_ids) do
+      local data = explorer.bookmarks[id]
+      local desc = data.desc or (vim.is_callable(data.path) and data.path() or data.path)
+      table.insert(lines, id .. ' │ ' .. desc)
+    end
+    table.insert(lines, '')
+  end
+
   table.insert(lines, '(Press `q` to close)')
 
   -- Create buffer
@@ -2095,6 +2166,27 @@ H.buffer_make_mappings = function(buf_id, mappings)
     return [[<C-\><C-n>]]
   end
 
+  local mark_goto = function()
+    local id = H.getcharstr()
+    if id == nil then return end
+    local data = MiniFiles.get_explorer_state().bookmarks[id]
+    if data == nil then return H.notify('No bookmark with id ' .. vim.inspect(id), 'WARN') end
+
+    local path = data.path
+    if vim.is_callable(path) then path = path() end
+    local is_valid_path = type(path) == 'string' and H.fs_get_type(vim.fn.expand(path)) == 'directory'
+    if not is_valid_path then return H.notify('Bookmark path should be a valid path to directory', 'WARN') end
+    MiniFiles.set_branch({ path })
+  end
+
+  local mark_set = function()
+    local id = H.getcharstr()
+    if id == nil then return end
+    local state = MiniFiles.get_explorer_state()
+    MiniFiles.set_bookmark(id, state.branch[state.depth_focus])
+    H.notify('Bookmark ' .. vim.inspect(id) .. ' is set', 'INFO')
+  end
+
   local buf_map = function(mode, lhs, rhs, desc)
     -- Use `nowait` to account for non-buffer mappings starting with `lhs`
     H.map(mode, lhs, rhs, { buffer = buf_id, desc = desc, nowait = true })
@@ -2106,6 +2198,8 @@ H.buffer_make_mappings = function(buf_id, mappings)
   buf_map('n', mappings.go_in_plus,  go_in_plus,            'Go in entry plus')
   buf_map('n', mappings.go_out,      go_out_with_count,     'Go out of directory')
   buf_map('n', mappings.go_out_plus, go_out_plus,           'Go out of directory plus')
+  buf_map('n', mappings.mark_goto,   mark_goto,             'Go to bookmark')
+  buf_map('n', mappings.mark_set,    mark_set,              'Set bookmark')
   buf_map('n', mappings.reset,       MiniFiles.reset,       'Reset')
   buf_map('n', mappings.reveal_cwd,  MiniFiles.reveal_cwd,  'Reveal cwd')
   buf_map('n', mappings.show_help,   MiniFiles.show_help,   'Show Help')
@@ -2797,6 +2891,12 @@ H.get_first_valid_normal_window = function()
   for _, win_id in ipairs(vim.api.nvim_tabpage_list_wins(0)) do
     if vim.api.nvim_win_get_config(win_id).relative == '' then return win_id end
   end
+end
+
+H.getcharstr = function()
+  local ok, char = pcall(vim.fn.getcharstr)
+  if not ok or char == '\27' or char == '' then return end
+  return char
 end
 
 -- TODO: Remove after compatibility with Neovim=0.9 is dropped